evaluation/requirements.txt
streamlit
python-dotenv
pdf2image
pytesseract
pymupdf
google-generativeai
